added missing console scipt
